FE-1470: Split the bindings' transport from the session facade
session.py mixed ~300 lines of process machinery with the 50-line
model-protocol facade a consumer actually reads. The machinery moves to a
private _transport.py (CliTransport: spawn, bootstrap, bounded reads,
timeouts, process-group shutdown, the JSON-lines exchange); the session
classes each own one transport and keep only their protocol vocabulary.

OptimizationSession gains from_manifest and from_manifest_file, mirroring
the base class's factories, so both classes construct the same way: name
the source, get the session. Both class docstrings now state the superset
rule — a manifest-serving process still answers every model-protocol
method, and only it answers optimization.*.

No public API changes; the suites run unchanged apart from three
assertions that reached into _request_timeout_seconds and two module-level
monkeypatches, which follow the constants to _transport.
